Moving ScreenLock to GUI
